swagger json location changed
