+++ src/app/layout.tsx
@@ -1,21 +1,5 @@
-import { Crimson_Text, Roboto } from "next/font/google";
 import "./globals.css";
-import Providers from "./providers";
 import { generateMetadata, viewport } from "./metadata";
-
-const crimsonText = Crimson_Text({
-  weight: ["400", "600", "700"],
-  style: ["normal", "italic"],
-  subsets: ["latin"],
-  variable: "--font-crimson",
-});
-
-const roboto = Roboto({
-  subsets: ["latin"],
-  weight: ["400", "700"],
-  style: ["normal", "italic"],
-  display: "swap",
-});
 
 export const metadata = generateMetadata();
 export { viewport };
@@ -27,9 +11,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en" suppressHydrationWarning>
-      <body className={`${crimsonText.variable} ${roboto.className} antialiased`}>
-        <Providers>{children}</Providers>
-      </body>
+      {children}
     </html>
   );
 }
